+++ include/dem/gear3_integrator.h
@@ -112,8 +112,7 @@ public:
             std::vector<Tensor<1, 3>>                       &force,
             const std::vector<double>                       &MOI,
             const parallel::distributed::Triangulation<dim> &triangulation,
-            typename DEM::dem_data_structures<dim>::cell_index_int_map
-              &cell_mobility_status_map) override;
+            AdaptiveSparseContacts<dim> &sparse_contacts_object) override;
 
 private:
   Point<dim>     predicted_location;