+++ include/dem/integrator.h
@@ -15,9 +15,9 @@
  *
  */
 
+#include <dem/adaptive_sparse_contacts.h>
 #include <dem/data_containers.h>
 #include <dem/dem_solver_parameters.h>
-#include <dem/disable_contacts.h>
 
 #include <deal.II/particles/particle_handler.h>
 
@@ -94,8 +94,7 @@ public:
             std::vector<Tensor<1, 3>>                       &force,
             const std::vector<double>                       &MOI,
             const parallel::distributed::Triangulation<dim> &triangulation,
-            typename DEM::dem_data_structures<dim>::cell_index_int_map
-              &cell_mobility_status_map) = 0;
+            AdaptiveSparseContacts<dim> &sparse_contacts_object) = 0;
 };
 
 #endif /* integration_h */