+++ include/dem/particle_particle_broad_search.h
@@ -12,11 +12,10 @@
  * the top level of the Lethe distribution.
  *
  * ---------------------------------------------------------------------
- *
  */
 
+#include <dem/adaptive_sparse_contacts.h>
 #include <dem/data_containers.h>
-#include <dem/disable_contacts.h>
 #include <dem/find_boundary_cells_information.h>
 #include <dem/particle_particle_contact_info.h>
 
@@ -70,15 +69,15 @@ public:
    * @brief Finds a vector of pairs (particle_particle_candidates) which shows the
    * candidate particle-particle collision pairs. These collision pairs will be
    * used in the fine search to investigate if they are in contact or not.
-   * This version of the function is used when disabling particle contacts
-   * regards mobility is enable.
+   * This version of the function is used when adaptive sparse contacts regards
+   * mobility is enable.
    *
    * @param particle_handler The particle handler of particles in the broad
    * search
    * @param container_manager The container manager object that contains
    * containers to modify of contact pair candidates with other
    * containers with neighbors lists
-   * @param disable_contacts_object The object that contains the
+   * @param sparse_contacts_object The object that contains the
    * information about the mobility status of cells
    */
 
@@ -86,7 +85,7 @@ public:
   find_particle_particle_contact_pairs(
     dealii::Particles::ParticleHandler<dim> &particle_handler,
     DEMContactManager<dim>                  &container_manager,
-    const DisableContacts<dim>              &disable_contacts_object);
+    const AdaptiveSparseContacts<dim>       &sparse_contacts_object);
 
   /**
    * @brief Finds a vector of pairs (particle_particle_candidates) which contains the
@@ -109,15 +108,15 @@ public:
    * @brief Finds a vector of pairs (particle_particle_candidates) which contains the
    * candidate particle-particle collision pairs. These collision pairs will be
    * used in the fine search to investigate if they are in contact or not.
-   * This version of the function is used when disabling particle contacts
-   * regards mobility is enable.
+   * This version of the function is used when adaptive sparse contacts regards
+   * mobility is enable.
    *
    * @param particle_handler The particle handler of particles in the broad
    * search
    * @param container_manager The container manager object that contains
    * containers to modify of contact pair periodic candidates with other
    * containers with periodic neighbors lists
-   * @param disable_contacts_object The object that contains the
+   * @param sparse_contacts_object The object that contains the
    * information about the mobility status of cells
    */
 
@@ -125,7 +124,7 @@ public:
   find_particle_particle_periodic_contact_pairs(
     dealii::Particles::ParticleHandler<dim> &particle_handler,
     DEMContactManager<dim>                  &container_manager,
-    const DisableContacts<dim>              &disable_contacts_object);
+    const AdaptiveSparseContacts<dim>       &sparse_contacts_object);
 
 private:
   /**