+++ include/dem/particle_point_line_broad_search.h
@@ -15,8 +15,8 @@
  *
  */
 
+#include <dem/adaptive_sparse_contacts.h>
 #include <dem/data_containers.h>
-#include <dem/disable_contacts.h>
 
 #include <deal.II/distributed/tria.h>
 
@@ -72,8 +72,8 @@ public:
     const std::unordered_map<
       std::string,
       std::pair<typename Triangulation<dim>::active_cell_iterator, Point<dim>>>
-                               &boundary_cells_with_points,
-    const DisableContacts<dim> &disable_contacts_object);
+                                      &boundary_cells_with_points,
+    const AdaptiveSparseContacts<dim> &sparse_contacts_object);
 
   /**
    * Finds a map of tuples (tuple of particle and the locations of beginning and
@@ -105,8 +105,8 @@ public:
       std::string,
       std::tuple<typename Triangulation<dim>::active_cell_iterator,
                  Point<dim>,
-                 Point<dim>>>  &boundary_cells_with_lines,
-    const DisableContacts<dim> &disable_contacts_object);
+                 Point<dim>>>         &boundary_cells_with_lines,
+    const AdaptiveSparseContacts<dim> &sparse_contacts_object);
 };
 
 #endif /* particle_point_line_broad_search_h */