+++ source/dem/gear3_integrator.cc
@@ -77,7 +77,7 @@ for (auto particle = particle_handler.begin();
   */
 }
 
-// Gear 3 not implemented for disabling contacts
+// Gear 3 not implemented for adaptive sparse contacts
 template <int dim>
 void
 Gear3Integrator<dim>::integrate(
@@ -88,11 +88,10 @@ Gear3Integrator<dim>::integrate(
   std::vector<Tensor<1, 3>> & /* force */,
   const std::vector<double> & /* MOI */,
   const parallel::distributed::Triangulation<dim> & /* triangulation */,
-  typename DEM::dem_data_structures<dim>::cell_index_int_map &
-  /* cell_mobility_status_map */)
+  AdaptiveSparseContacts<dim> & /* disable_contacts_object */)
 {
   throw std::runtime_error(
-    "Disabiling particle contacts not supported with explicit Gear 3 integrator, use Verlet integrator.");
+    "Adaptive sparse contacts not supported with explicit Gear 3 integrator, use Verlet integrator.");
 }
 
 template class Gear3Integrator<2>;