+++ source/dem/particle_particle_broad_search.cc
@@ -137,7 +137,7 @@ void
 ParticleParticleBroadSearch<dim>::find_particle_particle_contact_pairs(
   dealii::Particles::ParticleHandler<dim> &particle_handler,
   DEMContactManager<dim>                  &container_manager,
-  const DisableContacts<dim>              &disable_contacts_object)
+  const AdaptiveSparseContacts<dim>       &sparse_contacts_object)
 {
   // Pre-fetch and clear containers
   auto &local_contact_pair_candidates =
@@ -162,12 +162,14 @@ ParticleParticleBroadSearch<dim>::find_particle_particle_contact_pairs(
       // The main cell & its mobility status
       auto cell_neighbor_iterator = cell_neighbor_list_iterator->begin();
       unsigned int main_cell_mobility_status =
-        disable_contacts_object.check_cell_mobility(*cell_neighbor_iterator);
+        sparse_contacts_object.check_cell_mobility(*cell_neighbor_iterator);
 
-      // If main cell has status "inactive", skip to next main cell
-      // No needs to check if the main cell has any particle after this
-      // step since empty cells have "inactive" mobility status and are skipped
-      if (main_cell_mobility_status == DisableContacts<dim>::inactive)
+      // If main cell has status inactive or advected, skip to next main
+      // cell No needs to check if the main cell has any particle after this
+      // step since empty cells have inactive or advected mobility status
+      // and are skipped
+      if (main_cell_mobility_status == AdaptiveSparseContacts<dim>::inactive ||
+          main_cell_mobility_status == AdaptiveSparseContacts<dim>::advected)
         continue;
 
       // Get particles in the main cell
@@ -176,9 +178,9 @@ ParticleParticleBroadSearch<dim>::find_particle_particle_contact_pairs(
           particle_handler.particles_in_cell(*cell_neighbor_iterator);
 
       // Store other particles in the main cell as contact candidates if
-      // main cell is mobile only (this is equivalent to when disabling
-      // particle contacts feature is not enabled)
-      if (main_cell_mobility_status == DisableContacts<dim>::mobile)
+      // main cell is mobile only (this is equivalent to when adaptive sparse
+      // contacts feature is not enabled)
+      if (main_cell_mobility_status == AdaptiveSparseContacts<dim>::mobile)
         {
           // Find local-local collision pairs in the main cell, 1st particle
           // iterator is skipped since the main particle will not be
@@ -201,17 +203,23 @@ ParticleParticleBroadSearch<dim>::find_particle_particle_contact_pairs(
         {
           // Get mobility status of current neighbor cell
           unsigned int neighbor_cell_mobility_status =
-            disable_contacts_object.check_cell_mobility(
-              *cell_neighbor_iterator);
+            sparse_contacts_object.check_cell_mobility(*cell_neighbor_iterator);
 
-          // No storing of particles as candidate if main cell is active,
-          // but neighbor is not mobile (active or inactive).
-          // Particle contacts between particles in 2 active cells or with
-          // particles in inactive cells are irrelevant.
-          // In this case, we skip this neighbor cell.
-          if (main_cell_mobility_status == DisableContacts<dim>::active &&
-              neighbor_cell_mobility_status != DisableContacts<dim>::mobile)
-            continue;
+          // No storing of particles as candidate if main cell has any active
+          // status, but neighbor is not mobile (active or inactive). Particle
+          // contacts between particles in 2 active cells or with particles in
+          // inactive cells are irrelevant. In this case, we skip this neighbor
+          // cell.
+          if (neighbor_cell_mobility_status !=
+              AdaptiveSparseContacts<dim>::mobile)
+            {
+              if (main_cell_mobility_status ==
+                  AdaptiveSparseContacts<dim>::static_active)
+                continue;
+              if (main_cell_mobility_status ==
+                  AdaptiveSparseContacts<dim>::advected_active)
+                continue;
+            }
 
           // Store particles in the neighbor cell as contact candidates
           typename Particles::ParticleHandler<dim>::particle_iterator_range
@@ -240,12 +248,14 @@ ParticleParticleBroadSearch<dim>::find_particle_particle_contact_pairs(
       // The main cell & its mobility status
       auto cell_neighbor_iterator = cell_neighbor_list_iterator->begin();
       unsigned int main_cell_mobility_status =
-        disable_contacts_object.check_cell_mobility(*cell_neighbor_iterator);
+        sparse_contacts_object.check_cell_mobility(*cell_neighbor_iterator);
 
-      // If main cell has status "inactive", skip to next main cell
-      // No needs to check if the main cell has any particle after this
-      // step since empty cells have "inactive" mobility status and are skipped
-      if (main_cell_mobility_status == DisableContacts<dim>::inactive)
+      // If main cell has status inactive or advected, skip to next main
+      // cell No needs to check if the main cell has any particle after this
+      // step since empty cells have inactive or advected mobility status and
+      // are skipped
+      if (main_cell_mobility_status == AdaptiveSparseContacts<dim>::inactive ||
+          main_cell_mobility_status == AdaptiveSparseContacts<dim>::advected)
         continue;
 
       // Particles in the main cell
@@ -260,14 +270,21 @@ ParticleParticleBroadSearch<dim>::find_particle_particle_contact_pairs(
            ++cell_neighbor_iterator)
         {
           unsigned int neighbor_cell_mobility_status =
-            disable_contacts_object.check_cell_mobility(
-              *cell_neighbor_iterator);
+            sparse_contacts_object.check_cell_mobility(*cell_neighbor_iterator);
 
           // No storing of particles as candidate if main cell is active
-          // but neighbor is not mobile (active or inactive)
-          if (main_cell_mobility_status == DisableContacts<dim>::active &&
-              neighbor_cell_mobility_status != DisableContacts<dim>::mobile)
-            continue;
+          // but neighbor is not mobile
+          if (neighbor_cell_mobility_status !=
+              AdaptiveSparseContacts<dim>::mobile)
+            {
+              if (main_cell_mobility_status ==
+                  AdaptiveSparseContacts<dim>::static_active)
+                continue;
+              if (main_cell_mobility_status ==
+                  AdaptiveSparseContacts<dim>::advected_active)
+                continue;
+            }
+
 
           // Defining iterator on ghost particles in the neighbor cells
           typename Particles::ParticleHandler<dim>::particle_iterator_range
@@ -472,7 +489,7 @@ void
 ParticleParticleBroadSearch<dim>::find_particle_particle_periodic_contact_pairs(
   dealii::Particles::ParticleHandler<dim> &particle_handler,
   DEMContactManager<dim>                  &container_manager,
-  const DisableContacts<dim>              &disable_contacts_object)
+  const AdaptiveSparseContacts<dim>       &sparse_contacts_object)
 {
   // Pre-fetch and clear containers
   auto &local_contact_pair_periodic_candidates =
@@ -503,13 +520,15 @@ ParticleParticleBroadSearch<dim>::find_particle_particle_periodic_contact_pairs(
       auto cell_periodic_neighbor_iterator =
         cell_periodic_neighbor_list_iterator->begin();
       unsigned int main_cell_mobility_status =
-        disable_contacts_object.check_cell_mobility(
+        sparse_contacts_object.check_cell_mobility(
           *cell_periodic_neighbor_iterator);
 
-      // If main cell has status "inactive", skip to next main cell
+      // If main cell has status inactive or advected, skip to next main cell
       // No needs to check if the main cell has any particle after this
-      // step since empty cells have "inactive" mobility status and are skipped
-      if (main_cell_mobility_status == DisableContacts<dim>::inactive)
+      // step since empty cells have inactive or advected mobility status
+      // and are skipped
+      if (main_cell_mobility_status == AdaptiveSparseContacts<dim>::inactive ||
+          main_cell_mobility_status == AdaptiveSparseContacts<dim>::advected)
         continue;
 
       // Particles in the main cell
@@ -525,13 +544,17 @@ ParticleParticleBroadSearch<dim>::find_particle_particle_periodic_contact_pairs(
            ++cell_periodic_neighbor_iterator)
         {
           unsigned int neighbor_cell_mobility_status =
-            disable_contacts_object.check_cell_mobility(
+            sparse_contacts_object.check_cell_mobility(
               *cell_periodic_neighbor_iterator);
 
           // No storing of particles if main cell is active but neighbor is
-          // not mobile (active or inactive)
-          if (main_cell_mobility_status == DisableContacts<dim>::active &&
-              neighbor_cell_mobility_status != DisableContacts<dim>::mobile)
+          // not mobile
+          if ((main_cell_mobility_status ==
+                 AdaptiveSparseContacts<dim>::static_active ||
+               main_cell_mobility_status ==
+                 AdaptiveSparseContacts<dim>::advected_active) &&
+              neighbor_cell_mobility_status !=
+                AdaptiveSparseContacts<dim>::mobile)
             continue;
 
           // Defining iterator on local particles in the periodic neighbor
@@ -569,13 +592,15 @@ ParticleParticleBroadSearch<dim>::find_particle_particle_periodic_contact_pairs(
       auto cell_periodic_neighbor_iterator =
         cell_periodic_neighbor_list_iterator->begin();
       unsigned int main_cell_mobility_status =
-        disable_contacts_object.check_cell_mobility(
+        sparse_contacts_object.check_cell_mobility(
           *cell_periodic_neighbor_iterator);
 
-      // If main cell has status "inactive", skip to next main cell
-      // No needs to check if the main cell has any particle after this
-      // step since empty cells have "inactive" mobility status and are skipped
-      if (main_cell_mobility_status == DisableContacts<dim>::inactive)
+      // If main cell has status inactive or advected, skip to next main
+      // cell No needs to check if the main cell has any particle after this
+      // step since empty cells have inactive or advected mobility status
+      // and are skipped
+      if (main_cell_mobility_status == AdaptiveSparseContacts<dim>::inactive ||
+          main_cell_mobility_status == AdaptiveSparseContacts<dim>::advected)
         continue;
 
       // Particles in the main cell
@@ -590,13 +615,17 @@ ParticleParticleBroadSearch<dim>::find_particle_particle_periodic_contact_pairs(
            ++cell_periodic_neighbor_iterator)
         {
           unsigned int neighbor_cell_mobility_status =
-            disable_contacts_object.check_cell_mobility(
+            sparse_contacts_object.check_cell_mobility(
               *cell_periodic_neighbor_iterator);
 
           // No storing of particles if main cell is active but neighbor is
-          // not mobile (active or inactive)
-          if (main_cell_mobility_status == DisableContacts<dim>::active &&
-              neighbor_cell_mobility_status != DisableContacts<dim>::mobile)
+          // not mobile
+          if ((main_cell_mobility_status ==
+                 AdaptiveSparseContacts<dim>::static_active ||
+               main_cell_mobility_status ==
+                 AdaptiveSparseContacts<dim>::advected_active) &&
+              neighbor_cell_mobility_status !=
+                AdaptiveSparseContacts<dim>::mobile)
             continue;
 
           // Defining iterator on ghost particles in the neighbor cells
@@ -631,13 +660,15 @@ ParticleParticleBroadSearch<dim>::find_particle_particle_periodic_contact_pairs(
       auto cell_periodic_neighbor_iterator =
         cell_periodic_neighbor_list_iterator->begin();
 
-      // If main cell has status "inactive", skip to next main cell
-      // No needs to check if the main cell has any particle after this
-      // step since empty cells have "inactive" mobility status and are skipped
+      // If main cell has status inactive or advected, skip to next main
+      // cell No needs to check if the main cell has any particle after this
+      // step since empty cells have inactive or advected mobility status
+      // and are skipped
       unsigned int main_cell_mobility_status =
-        disable_contacts_object.check_cell_mobility(
+        sparse_contacts_object.check_cell_mobility(
           *cell_periodic_neighbor_iterator);
-      if (main_cell_mobility_status == DisableContacts<dim>::inactive)
+      if (main_cell_mobility_status == AdaptiveSparseContacts<dim>::inactive ||
+          main_cell_mobility_status == AdaptiveSparseContacts<dim>::advected)
         continue;
 
       // Particles in the main cell
@@ -652,13 +683,17 @@ ParticleParticleBroadSearch<dim>::find_particle_particle_periodic_contact_pairs(
            ++cell_periodic_neighbor_iterator)
         {
           unsigned int neighbor_cell_mobility_status =
-            disable_contacts_object.check_cell_mobility(
+            sparse_contacts_object.check_cell_mobility(
               *cell_periodic_neighbor_iterator);
 
           // No storing of particles if main cell is active but neighbor is
-          // not mobile (active or mobile)
-          if (main_cell_mobility_status == DisableContacts<dim>::active &&
-              neighbor_cell_mobility_status != DisableContacts<dim>::mobile)
+          // not mobile
+          if ((main_cell_mobility_status ==
+                 AdaptiveSparseContacts<dim>::static_active ||
+               main_cell_mobility_status ==
+                 AdaptiveSparseContacts<dim>::advected_active) &&
+              neighbor_cell_mobility_status !=
+                AdaptiveSparseContacts<dim>::mobile)
             continue;
 
           // Defining iterator on local particles in the neighbor cells