+++ source/dem/periodic_boundaries_manipulator.cc
@@ -175,7 +175,7 @@ template <int dim>
 bool
 PeriodicBoundariesManipulator<dim>::execute_particles_displacement(
   const Particles::ParticleHandler<dim> &particle_handler,
-  typename DEM::dem_data_structures<dim>::periodic_boundaries_cells_info
+  const typename DEM::dem_data_structures<dim>::periodic_boundaries_cells_info
     &periodic_boundaries_cells_information)
 {
   bool particle_has_been_moved = false;